pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "boxoban-gym"
version = "0.1.0"
description = "Gymnasium environments for DeepMind Boxoban levels"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "Apache-2.0" }
authors = [
  { name = "boxoban-gym contributors" }
]
dependencies = [
  "gymnasium>=1.0",
  "numpy>=1.24",
]
keywords = ["reinforcement-learning", "gymnasium", "sokoban", "boxoban"]
classifiers = [
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "License :: OSI Approved :: Apache Software License",
  "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
]
wrappers = [
  "Pillow>=9.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
minversion = "8.0"
testpaths = ["tests"]
